Made Simulator support customizable speed
- Make Simulator support customizable speed and single-step execution (currently not testable because there is no UI support yet)
- Fix potential bug due to use of int8_t instead of uint8_t when reading/writing to save file
- Added MSVC-specific pragma to enable visual styles, so the dialog box buttons look like modern Windows buttons

+++ CircuitPlayground/filesaveaction.hpp
@@ -58,7 +58,7 @@ private:
                     },
                 }, element);
 
-                int8_t elementData = (element_index << 2) + (logicLevel << 1) + defaultLogicLevel;
+                uint8_t elementData = static_cast<uint8_t>((element_index << 2) + (logicLevel << 1) + defaultLogicLevel);
                 saveFile.write(reinterpret_cast<char*>(&elementData), 1);
             }
         }
